[FIX] stock_barcode: no suggest to scan lot/serial
Go to Inventory>Configuration>Operations Types
Select the <Current Company>: Receipts
  Uncheck Create New Lots/Serial Numbers
  Uncheck Use Existing Lots/Serial Numbers
  Check Show Detailed Operations
  Check Pre-fill Detailed Operations
Add a barcode on a [DEMO] prod, tracked by SN
Create a purchase order for [DEMO], confirm
Open barcode app to receive the product
Scan the barcode.

The system will prompt the user to
'Scan the serial or lot number of the product'
Even if creating serial numbers was disabled

Changes introduced with this commit:

* when scanning the product tracked by SN barcode each new line
create will have initial qty. 1 instead of 0.
* Hide the lot/serial in the barcode line when the user has configured
the picking type to not insert them at reception.
* Add a test for this use case.

opw-2423146

closes odoo/enterprise#16529

X-original-commit: f457be88feaaa1f5bb22032221869092201094a0
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>
Signed-off-by: Steve Van Essche <svs-odoo@users.noreply.github.com>