[FIX] l10n_mx_edi: duplicate cfdi document
Since, 06a82a94fc4310dcb9b0570717338891cfaaf33b when an error occurs
during the sign process a xml is also created, this was done to allow a
better debug and understand of the issue. The issue, is that each time
that there was an error a new xml was created.

Now, there is only one cfdi document created.

opw-2440256

closes odoo/enterprise#15885

X-original-commit: 0df2003d4a6018045ded3bc8864528ad88a81605
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>