[FIX] l10n_mx_edi: finkok error with signed document
Before this commit, if finkok (Quadrum) answers with an error along with
the signed document, the xml on Odoo will not be signed and only the
error is shown. This occurs for example with  : Error 307 : 'El CFDI
contiene un timbre previo'.

Now, if there is an error along the signed document, the error will not
show, and the document will be signed on Odoo.

opw-2440256

closes odoo/enterprise#15884

X-original-commit: 7d20306c59f484b60376d142177af22689affba3
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>