[FIX] account_asset: fix computed depreciated amount when selling asset
- Go to Accounting > Accounting > Assets
- Create an Asset:
  * Original Value: $ 100
  * Acquisition Date: 31/12/2019
  * Method: Linear
  * Duration: 8 Months
  * First Depreciation Date: 31/01/2020
- Confirm
- Click on all 8 Depreciation Lines to post them
- Create an Invoice of $ 110 with Invoice Date set to 05/05/2020 and post it
- On created Asset page, sell it by selecting the created Invoice
- You will need to reverse the 4 last Depreciation Lines as their date are after Invoice Date
- Once the 4 lines reversed, sell the Asset
- Check the value of the Current Depreciation line of the Sale move (i.e. Value A)
- Post the Sale move and check Depreciable Value and Book Value of Asset
If we follow the exact same steps without posting the 4 last Depreciation Lines (which means
that we will not reverse them), Value A, final Depreciable Value and Book Value will be different.
But the values should be the same.

When selling the Asset, the depreciated amount is computed from all posted depreciation lines,
but it should be computed from posted depreciation lines that have not been reversed.

opw-2390385

closes odoo/enterprise#16018

X-original-commit: 7171d20d89b73a9fe88ac247635c69b59487a611
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>