[FIX] hr_contract_salary: Fix the traceback on customize salary
When a floating, string(blank) value is entered in the integer field in the customize
your salary, it will be given a traceback for invalid value.

For Ex:
    1) when user change km_home_work from salary configurator it will
    give the traceback.

    2) when user removing the value of km_home_work after 'TAB' key press
    from salary configurator it will be given the tracebak.

So in this commit, on change and on submit of the salary configuator
check for valid value and if value is not valid then display validation
error to user.

closes odoo/enterprise#16754

Taskid: 2410285
X-original-commit: be1ab29ba1f1aecb7370dfdd19ae7f4156c3b4d2
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>