[FIX] sale_subscription_dashboard: Net revenue
When checking the revenue KPIs, the net revenue is always 0.

To reproduce the error:
1. Go to sales, create a SO
	- Product must be a subscription product
2. Confirm, Create Invoice
3. Post the invoice, Register the payment
4. Go to Subscriptions
	- You find the subscription you just created in the In-Progress
column
5. Reporting > Revenue KPIs
6. Set the date range so your subscription will be included, Update

Error: All charts but the Net Revenue are updated. The Net Revenue is
still $0. It should at least include the sold and paid subscription
(from steps 1-3).

The error comes from the SQL request conditions used to compute the net
revenue. The SQL result includes the reconciliation lines, therefore the
sum is 0.

OPW-2431880

closes odoo/enterprise#15987

X-original-commit: 532757ab1c99b52594f4b8d1f1abdd7172e08aa2
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>