[ADD] l10n_ar_edi: add QR code to invoice report
This is change is required by AFIP RG 4291 that will start been mandatory from
March 1st 2021. This one make mandatory to show a QR code in the printed
version of the electronic invoices. We also make this improves:

1) QR code replace the old barcode
2) this code will be printed only for electronic and online invoices,
formerly it was wrongly added also in the vendor bills.
3) translations were updated to add the new QR code field

NOTE: Fix pylint errors.  This ones pop up when running this PR "Check Style"
in runbot

closes odoo/enterprise#15949

Related: odoo/odoo#65034
Signed-off-by: Josse Colpaert <jco@openerp.com>