[FIX] stock_barcode: separate code/display name
Since 3c7f6bde6753da15fd13007ed73ddd727fad0f8b, on the barcode line, the
product code is displayed on top of the product name.
But if a product is added by scanning a product non-present yet, the
added line has the product display name in one line, instead of having
its code and product name on different line.

Also, for an inventory, the product's code and name are always on the
same line. As it shouldn't be different, fixed that too.

task-2447576

X-original-commit: ab2180f2e0be955036f8e7975a7e6badd04df520