[REF] stock_barcode: common code between pick/inv
Some part of codes are the same in inventory side than picking side.

  - In the method `get_barcode_view_state`, the part to get `barcode`,
`tracking` and now `code` are the same both for `stock.picking` and
`stock.inventory`. Moves this part in `product.product` in the model's
method `_get_fields_per_product_id`.

  - The method `_makeNewLine` in client actions:
This method was abstract in the abstract client action and returned
nothing but an empty object.
As there is a lot of common attributes for the inventory and picking
lines, the abstract model can instead return an object with the common
attributes, then specific client actions can add their own attributes.

Also, removes the `reference` attribute as it was never used (and was
always `undefined` anyway).

X-original-commit: 1d5e965dbbcc2eae793b2ba3f25ef8b72241798d