[FIX] mrp_workorder: additional components used twice in MO
When adding components to a confirmed MO in tablet view, then not
processing the full quantity at once and creating a backorder
for the remaining quantity, the consumption of the added components
in the first backorder is not correct.
There was a duplication of the move line of the added components,
doubling its used quantity. This duplication came from two situations:
 - If the additional component is available, there was no move line in the
quality check created with the additional component. Not seeing this move
line, a new one was created in the _next method of mrp_workorder.
 - If the compoment is not available, no SML is recorded in the quality
check but a SML was created in set_qty_producing because no workcenter
was registered in the stock move. A second SML was then created in _next.

closes odoo/enterprise#19028

Task-id: 2447949
X-original-commit: 980ccaacb7a6b5dd99246ee65317c0a80f1db8cb
Signed-off-by: Rémy Voet <ryv-odoo@users.noreply.github.com>
Signed-off-by: npi-odoo <npi-odoo@users.noreply.github.com>