[FIX] industry_fsm_stock: fix add new product in sale order from fsm task
Currently, when we create SO with FSM product and Inventory
product and confirm it after that when we update the qty or
add new product from the FSM task then it will neither update the
qty on Delivery order or nor Generate an exception on devliery order.

So in this commit, when the user marks the FSM task as done, then
the system will generate the exception activity for the delivery order
based on qty update on SO line.
Also if there is exception do not validate the delivery order it will
be validated manualy by the responsible user.

TaskID:2370887

closes odoo/enterprise#16065

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>