[FIX] account_bank_statement_import_csv: redirection after import
When importing some bank statements, the page is redirected to the
reconciliation page. However, if the user then tries to reconcile an
imported bank statement line, an error will be displayed: the line must
be posted first. Indeed, since version 14, a bank statement needs to be
posted.

Therefore, this fix changes the redirection page: it opens the new bank
statement so the user can post it and then reconcile each line.

OPW-2444118

closes odoo/enterprise#16169

X-original-commit: ab0f8456711f7aefe6f6c695ed6fa5acc67a29ea
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>