[ADD] l10n_de_pos_cert + l10n_de_pos_res_cert for Germany POS localization
According to the German regulations, every PoS (restaurant included) need to use a Technical Security System (TSS). A cloud-based solution has been opted with Fiskaly which provides the four legal requirements:
  - (1) Registration of the company at the financial authority
  - (2) TSS for the PoS
  - (3) DSFinV-K
  - (4) Archiving

(2) The cloud TSS records in real time the order information (only for restaurant) and payment details during a PoS session.

(3) The DSFinV-K is a service used to store all the order information when closing a session. An export of those information can be downloaded in case of an audit.

closes odoo/enterprise#18520

Task-id: 2055957
X-original-commit: d2d2f1bd0792ec9f0ed7eff43b5cc90a5dabcbb0
Related: odoo/odoo#71247
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>
Signed-off-by: Olivier Dony (odo) <odo@openerp.com>