[FIX] web_grid: fix post-cell-edit glitch
This commit fixes the post-cell-edit glitch issue by updating value in
the model state before rpc call.
This commit adds a test to reproduce the glitch.

------------------------------------------------------------

The issue can be described as follow : the edited value
disappear during a while when the user has updated a value
and is then re-rendered. On large databases, the glitch can
last a longer time and is more visible.

Previous behaviour :

The previous value of the update cell is shown the time between the
two rpc calls (adjust_grid, read_group) because the rendering is done
just after adjust_grid with editMode not set and the cell value not
updated in the model rendered.

    - User triggers _onUpdateValue by writing a new value
      (let's say : (new : 42, old : 0))
    - The rendering is updated during the call to rpc in the
      grid_controller.js _adjust method.
         -> As the state.editMode to render the web_grid.CellValue,
            the rendering is a div with the cell.value user, which is
            the previous state value : 0.
    - The rendering is updated once again after the result of read_group
      => Cell value = new_value : 42.

Expected Behaviour :

Between the two rpc calls, the editMode is still true, which allows the
web_grid.CellValue template to render an input cell with the
currentInput value : 42. The user get not troubled by the value changes
and the behaviour is more smooth regarding UX.

------------------------------------------------------------

task-2389731

closes odoo/enterprise#16329

Closes: #16006
X-original-commit: 0d8b7c85eafb67f598ba8be26936d1687d06563a
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>