[FIX] delivery_easypost: Tracking Reference is usable without sending
Steps:
- Go to Sales > Configuration > Sales Orders / Shipping Methods
- Create
  - Name: (1)
  - Provider: Easypost
  - Delivery Product:
    - Name: Easypost
    - Product Category: All / Deliveries
    - Product Type: Service
  - Enter your API keys
  - Click "Load your Easypost carrier accounts"
  - Select FedEx
  - Default Package Type: FexExTube
- Go to Orders > Quotations
- Create a quotation with one product
- Confirm
- Click the "Delivery" smart button
- Edit
  - Additional Info tab:
    - Carrier: (1)
    - Tracking Reference: Any valid tracking code
- Click the "Tracking" smart button

Bug:
Error: Easypost returned an error: The requested resource could not be
found.

Explanation:
When entering the tracking reference manually, the Easypost order ID is
not present. This ID is only added when shipping through Odoo. However,
we can see that the Tracking smart button is still appearing when saving
the `stock.picking`. Clicking this smart button calls Easypost on the
path `/orders/False`. This returns an error.

This commit calls the correct route, which is `/trackers`, when there is
no Easypost order ID on the `stock.picking`.
The fix also takes into account the fact that there could be multiple
tracking references in this field. These are separated by `+` signs as
seen here:
https://github.com/odoo/enterprise/blob/616bc667b52ea6b0a3b5c304dc59c05f0c9b7fb1/delivery_easypost/models/delivery_carrier.py#L99

opw:2439976

closes odoo/enterprise#16153

X-original-commit: 6cc637f00222a20a7d22d26eff02e59f0f66f37d
Signed-off-by: backspac <backspac@users.noreply.github.com>