[FIX] l10n_pe_edi, l10n_mx_edi: remove useless ormcache invalidation
The ormcached methods do not need invalidation. Their cache entries depend on
the value of the fields of the input record, not on the record itself. And as a
cache value only depends from its cache key, the cache entries are simply never
invalid. In other words, modifying a record does not make the cache invalid.

Therefore, the ormcache invalidation is useless. And it is even harmful in this
case, as it clears the whole cache, which implies some performance loss.

closes odoo/enterprise#16017

closes odoo/enterprise#16094

X-original-commit: 6cd9fb8b7d83fe6d3a99418f0d751bb16c1e6d52
Signed-off-by: Raphael Collet (rco) <rco@openerp.com>
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Josse Colpaert <jco@openerp.com>