[FIX] account_reports: mounted called twice
Before this commit, the 'mounted' method of the ControlPanel in
the account_reports was called twice.

It happened because the account_reports updated the ControlPanel before
being actually mounted, so mounted was called once when the
account_reports was mounted, and once when the update was applied.

Ideally, this should not be an issue (this isn't an issue with owl).
However, in Odoo, we mix layers of Owl Components and legacy
widgets. In these situations, the above scenario isn't properly
handled (and can't be).

As a consequence, in mobile (enterprise), it crashed because an
handler bound in mounted (thus twice) was only unbound once.

This commit avoids the issue as the update was actually useless.

Steps to reproduces (Mobile):
* Go to Accounting
* Go to "Balance in GL" (Bank card)
* Unfold the 101402 Bank line
* Open the dropdown of one sub line (e.g. BNK1/2021/01/0001)
* Select view in the dropdown
* Go back to the previous view using breadcrumb
* Optional: Go to another app if the screen can't scroll (e.g. go to Sales)
* Scroll the view => Bug

closes odoo/enterprise#16507

X-original-commit: d82e4f6b02baa8dff3396604fd59e69c4f77993c
Related: odoo/odoo#66497
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>