[IMP] industry_fsm: move the demo records from 'done' stage to 'new'
Currently, the some task are in 'done' stage and marked as done but
they are in future dates or today.

So in this commit, move the tasks which are in done stage and
planned for today or future to the 'new' stage and do not mark them
as Done.

TaskID : 2445906

closes odoo/enterprise#16046

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>