[FIX] account_online_sync: allow to delete user from duplicated db
Two instances (database) of Odoo can't run with the same connection
towards a banking institution. In case such a scenario happens,
The duplicated database will receive an error.
The problem is that users can't remove their connections on such db
as an error is raised even when trying to delete the duplicated
connection. This commit fixes this problem by ignoring
- signature errors
- account not found errors
- provider data corrupted errors
when trying to delete a connection. This will allow the user to remove
their record from their database and will keep the existing information
on the proxy.

closes odoo/enterprise#16064

X-original-commit: 5bbbe392fe74f3ee35a4e4b027f2dc8008d15520
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>