[FIX] account_bank_statement_import_camt: select correct rate
When importing a CAMT bank statement, if the latter uses several
currencies, the module may convert amounts unnecessarily.

To reproduce the error, import this bank statement:
[2439413_CAMT_example.xml](https://drive.google.com/file/d/1W4kAoZ4RQkrp1GjrLMSbwP_4BrQIpL-w/view?usp=sharing)
Once done, the user is redirected to the reconciliation page and has to
reconcile $1000. However, here is an extract of the CAMT bank statement
XML:
```xml
      <Ntry>
        <Amt Ccy="USD">500.00</Amt>
        <CdtDbtInd>CRDT</CdtDbtInd>
        <Sts>BOOK</Sts>
        <AmtDtls>
          <TxAmt>
              <Amt Ccy="EUR">250</Amt>
              <CcyXchg>
                  <SrcCcy>EUR</SrcCcy>
                  <TrgtCcy>USD</TrgtCcy>
                  <XchgRate>2</XchgRate>
              </CcyXchg>
          </TxAmt>
        </AmtDtls>
	...
      </Ntry>
```
The original amount was 250€. Thanks to the change rate (2), the final
amount is $500. Therefore, the user should reconcile $500, not $1000.

For now, here is how the module calculates the amount of an entry:
```python
_get_amount = partial(_generic_get,
    xpath='ns:Amt/text() | ns:AmtDtls/ns:TxAmt/ns:Amt/text()')

_get_rate = partial(_generic_get,
    xpath='ns:XchgRate/text() | ns:AmtDtls/ns:TxAmt/ns:CcyXchg/ns:XchgRate/text()')

def _get_signed_amount(*nodes, namespaces):
    amount = float(_get_amount(*nodes, namespaces=namespaces))
    rate = float(_get_rate(*nodes, namespaces=namespaces)) or 1.0
    sign = _get_credit_debit_indicator(*nodes, namespaces=namespaces)
    return amount * rate if sign == 'CRDT' else -amount * rate
```
(From account_bank_statement_import_camt/wizard/account_bank_statement_import_camt.py)

`_get_signed_amount` and the XPaths need to be updated. In the above
case, `amount` is defined thanks to the first `<Amt>` node ($500).
However, `rate` is also defined (2). As a result, the module applies the
conversion again (this is the reason why the user has to reconcile $1000
instead of $500).

This fix first tries to use the data inside `<AmtDtls>` node ("Amount
Details"). If there isn't any amount, it uses the data directly present in
`Ntry` node. It then applies the conversion only if the amount currency
is different from the current of the target conversion (`<TrgtCcy>`).

OPW-2439413

closes odoo/enterprise#16443

X-original-commit: a54d9d562a55a94c5f5fa218559cc89ba12d2a1d
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>