[FIX] industry_fsm_sale: use the button define in kanban view
Before this commit, we are looking for the oe_kanban_action
containing the '+' but now this button is integrated in the widget and
thus it is no longer a oe_kanban_action.

This commit uses the oe_kanban_action to add a quantity for the product
when the user clicks on the kanban record.

task-2447258