[FIX] web_map: misaligned record label in view
Before this commit, the label of unmapped records is displayed below the number
because of d-flex class used on div.

This commit replaces the uses d-inline class to display number and label of
unmapped record on same line, also added title attribute so if label is too
long then help title is displayed.

task-2416755

closes odoo/enterprise#19503

X-original-commit: 8df25564404ee0a040af658b1318b6fad017e786
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>
Co-authored-by: Mohammed Shekha <msh@odoo.com>