[FIX] account_asset: Multi Company Enabled in Asset Report
Beforehand, the assets report could only show detailed data from
one single company. If the user was logged into several companies
at the same time, a "no account" line would appear in the report,
that would contain the data of the secondary companies the user
would be logged to.

A first PR (https://github.com/odoo/enterprise/pull/15989) was made
to fix the issue, but it introduced a new bug.
If 2 accounts accross differents companies have the code, then a
stacktrace would appear.

Now, all the different accounts accross the potentially multiple
companies are handled correctly, and the details of all the assets
are displayed for alle the companies the user is logged into.

closes odoo/enterprise#16326

X-original-commit: 36d066e84700ef37bc86ed931ac728dd5709fd44
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Julien CHEVREAU <Julien-CHEVREAU@users.noreply.github.com>