[FIX] account_invoice_extract: warning when the invoice date is before lock date
When a lock date has been set, it could prevent the extraction system from updating.
There is now a warning displayed and the date is set to the next day following the
lock date.

This required a modification to the way the warnings are represented. As multiple
warnings could occur at the same time on an invoice, simple codes for each of them
was insufficient.
Instead, they're now represented by power of 2's so that they can be combined by
summing them. Analyzing the bits of the resulting number allows to retrieve which
warning are active.

closes odoo/enterprise#16982

X-original-commit: e2a8b915558c34fad38f303e5cf5a29a4fb6a6a0
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>