[IMP] l10n_be_hr_payroll: Add NISS field on employee
Purpose
=======

All the employees (including foreigners) have a NISS number, given at
birth for belgian, of from the DIMONA declaration for foreigners.

This is the number that is used to identify someone when declaring
withholding taxes to the state (or meal vouchers to the provider).

As we also need the national number of foreigners in their own country,
we add a field that will be used in the DMFA and the 273S sheets.