[FIX] mail_enterprise: fix test to disable back button
This commit fixes a test in chat window manager to make it work
with the changes introduced by PR 64599 on Odoo.

The messaging menu needs to be re-open when a chat window is closed,
so we need to close it because it overrides the back button too.
As long as something overrides the back button, it can't be disabled.

closes odoo/enterprise#16323

Tasks-id: 2267593
X-original-commit: c2b27648ec3ec7f05a4d58035cc133ae0882c29e
Related: odoo/odoo#66102
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>