[IMP] industry_fsm_sale: generate SOLs in SO for employee rate
When the pricing type of a fsm project is equal to employee_rate and the
user wants to mark as done a task into this project. Then we need to
generate the SO (if needed) and its SOLs to link to the task.

This commit manages when the employee is in the employee mappings, we
generate a SOL if the product and the price unit of it is not in the SO.
Then the link is done to the corresponding timesheets.

task-2376382

Community PR: odoo/odoo#66157
Upgrade PR: odoo/upgrade#2160

closes odoo/enterprise#16348