[IMP] industry_fsm_sale: determine the correct SOL for timesheet
When the user changes the SOL in task for the one in another SO,
we need to recompute the SOL in each timesheet of this task.
If the pricing type of the fsm project is employee rate then we
search if there is a mapping for the employee who does the timesheet. If
yes, then we search an SOL which has the product and the same price unit
than the mapping found. Otherwise, if no mapping or no sol is found then
we give the SOL of the task.

task-2376382

Community PR: odoo/odoo#66157
Upgrade PR: odoo/upgrade#2160

closes odoo/enterprise#16348