[IMP] industry_fsm_sale: click on kanban record to add quantity
Before this commit, a invisible button is created in the kanban view to
simulate a click on the kanban record to add an unit in the product
quantity. The problem with this method is we do not keep the value
edited in the product quantity but we add an unit to the quantity stored
in the backend.

This commit removes this button in the kanban view and add a click event
in the kanban record, this event is triggered when the user clicks on
the kanban record of the product and it calls the method in the
FSMProductQuantity to add an unit to the product quantity.

task-2447258