[IMP] website_calendar: redirect internal user to event form view
Right now, if the attendee clicks on the 'View' button from the confirmation
mail received after attendee books online apointment, the responsible gets
redirected to the website view of the event. However, if the attendee is
an internal user, he should land on the form view of the event instead of
seeing the website page for this event.

This commit improves the behavior by redirecting attendee to backend view
of the event in case the attendee is internal user.

Task ID-2199620
COM PR odoo/odoo#64925
ENT PR odoo/enterprise#15914