[FIX] account_online_synchronization: fix balance being wrongly updated
It can happen that we receive no new transactions but that the balance
of the account has already been updated. In which case the balance
would write itself in the previous existing statement in Odoo which we
don't want. -> Ignore writing balance on the latest statement if no
new transactions have been fetched.
Also add a chatter message whenever we reset a statement to draft, not
just when the statement was in done mode.

closes odoo/enterprise#16256

X-original-commit: acf26c07733d528e2e1b225e0f8255874cf68569
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>