[REF] website_event_track_social: split computed editable fields
Purpose
=======
Some computed stored editable fields are computed in the same method.

In that case, when one of the value is provided in a create / write call,
the computed method is not called and other fields computed in the same
method do not have the right value.

Technical
=========
We use a post init script to be sure to not send push notification for
existing event track. So if we first install the Event module and website,
then activate push notification and the install the module
"Website Event Track Social" we won't send notification for old event
tracks.

Task 2377119
COM odoo/odoo/pull/65772
ENT odoo/enterprise/pull/16221

X-original-commit: 1e01a8f75046be50d92a73f0e919495c74ef97bd