[IMP] industry_fsm_sale: add warning message if SOL in task changed
Before this commit, when the user wants to change the SOL in fsm task
for a SOL from another SO then the SOL linked to the timesheets in
this task is also impacted of this changed.

This commit informs the user the impact of the SOL changes in a fsm
task.

task-2376382

Community PR: odoo/odoo#66157
Upgrade PR: odoo/upgrade#2160

closes odoo/enterprise#16348