[IMP] payment_sepa_direct_debit: better error message when reusing a res.partner.bank
When paying a web order using SEPA Direct Debit, if the IBAN account entered by the user already existed in the database, an error message was displayed to the user, directly telling an account only can belong to one partner ( https://github.com/odoo/odoo/blob/14.0/odoo/addons/base/models/res_bank.py#L92 ) . This was pretty obscure, so we replace it by something more user-friendly. We ask the user to login, as it is very likely the error occurs simply because he already had an account and is recreating it entirely instead of reusing his existing account.

X-original-commit: 7649b0baa47895278bb5d4f56d7e440429e76391