[FIX] account_assets_report: multi-currency on deprecation schedule report
Before this commit, if an asset was encoded on a foreign currency
(example, euros if the company currency is dollars). On the deprecation
schedule report, the values shown was the ones of the foreign currency
with the company currency logo.

Now, all the foreign currency assets are convert into company currency
on the acquisition date.

opw-2451383

closes odoo/enterprise#16648

X-original-commit: 732d2fba41919ff055944baad2f69345775b176c
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>