[IMP] sign: display a bottom sheet to fill in text fields on mobile
Before this commit, text fields were very small and it's was hard to see
what we were writing in it. The only possibility was to zoom in the PDF.

In this commit, a bottom sheet with a text field in it now appears when
you start to fill in the form. you clearly see what you are writing.
Note that the zoom is still essential if you want to see the pdf
content.

Task ID: 2463472