[IMP] industry_fsm_sale: reduce font size if more 5 digit and limit to 10
Before this commit, the user can enter more than 10 digits but the
parseValue can return not valid integer because we can have an overflow.

Thus, this commit limits the number of digits of the fsm quantity to 10.
Moreover, when the user enters more than 5 digits than the font size is
reduced a little bit to prevent to resize the kanban record.

task-2447258