[FIX] documents_spreadsheet: get archived related records
In the pivot view, group projects by customer. Even if a customer is archived,
it will appear in the pivot view since the project itself is not archived.

When this pivot is inserted in a spreadsheet, the group with this customer
won't appear because we perfom a distinct `search_read` to order groups
(based on group values). If we group by a relational field, we must ensure
we retrieve the record, event if it's archived.

opw 2454906

closes odoo/enterprise#16447

X-original-commit: abb25826d2ba6b534242550ba7589e76f2d8a126
Related: odoo/odoo#66393
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>