[FIX] web_studio: keep 'existing field' unfolded in kanban and search
before this commit: the kanban view and search view contain only the
'existing field' section and it is folded by default because of this behavior,
users have to unfold the 'existing field' section every time even if it is the
only content in the 'add' tab.

with this commit, we fixes the issue by making the 'existing field' section
unfolded by default in kanban view and search view.

task-2453573

closes odoo/enterprise#18565

X-original-commit: fe58c3673cb3b221f1b19a029aee7ae370aa40f5
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>