[IMP] industry_fsm_sale: compute pricing_type when first mapping created
Before this commit, when the user creates the first mapping in the fsm
project, the compute for the pricing_type is not triggered to set its
value from 'task_rate' to 'employee_rate'.

This commit override the create method of 'project.sale.line.employee.map'
model to update the pricing_type of its fsm project if the value is not
the one expected.

task-2376382

Community PR: odoo/odoo#66157
Upgrade PR: odoo/upgrade#2160

closes odoo/enterprise#16348