[FIX] l10n_pe_edi: avoid problems at module installation
Before, when you had already a Peruvian company with posted_before
customer invoices, the installation of the module would fail in
the post_init_hook as it would not be able to let it set
l10n_latam_use_documents on the journal.  Now we check this first.

Also in the function that checks if the invoicing is required, we
check if the journal uses document types/numbers, so the old journal
can still exist.

closes odoo/enterprise#16369

X-original-commit: 527386e0f3998f5e63c5662a3985d565c8fed7b4
Signed-off-by: Josse Colpaert <jco@openerp.com>