[IMP] industry_fsm_sale: add timesheet products based on employee rate
Before this commit, when the pricing type of the fsm project is employee
rate and the user marked as done a task in this project. The SOLs in the
SO (generated also if it is not the case) are generated to add the
service products defined in employee mapping and the default service
product in the project. The problem is the service products coming from
the mapping are considered as additional materials for the task.

This commit adds a condition to consider these products as well as
service products and not as additional materials for the tasks.

task-2376382

Community PR: odoo/odoo#66157
Upgrade PR: odoo/upgrade#2160

closes odoo/enterprise#16348