[IMP] industry_fsm_sale: simplify pricing type in project settings
Now, the pricing_type field in the project will be a compute non-stored
field and this field will not be visible in the UI.
This commit updates the condition on the fields visibility was impacted
by this field. Because the pricing_type field is now computed based on
these fields, that is based on the configuration of the project.

task-2442683

Community PR: odoo/odoo#66181
Upgrade PR: odoo/upgrade#2164
closes odoo/enterprise#16356