[FIX] helpdesk: set ticket close date according to state
Currently, When user move the ticket the close(solved) stage
the close Date of the ticket will be added in close_date field.
But when user move the ticket from closing stage to other stage
at that time close date is not updated.

So in this commint, When change the stage of tickets from the closing
stage to other stage update the Close date to False.

closes odoo/enterprise#16460

Taskid: 2379017
X-original-commit: 6a27ceb906dc72bdb5b21682d4036a90775b9b83
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>