[FIX] voip: traceback when hanging up the call
Before this commit:

An activity "call" is already marked as "done" in chatter, after that user
schedule an activity "call" again and actually make that call and hang up the
call, traceback will be occured. And another issue is when page will be
refreshed, this activity will be marked as "done" and appears in chatter.

After this commit:

Hangs up the call for the second time, the activity will be marked as
"done", and appear in the chatter with the original notes ( in "Today"
activities section ) without any traceback.

Links

PR https://github.com/odoo/enterprise/pull/16464
Task- 2457219

closes odoo/enterprise#16683

X-original-commit: 56944bad19d771e8c9d6a9d668042f150cefcd50
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>