[FIX] account_asset: Generate Disposal Move for fully Depreciated Assets
Beforehand, when disposing an asset, a disposal move was generated only
if the asset still had draft depreciation moves, that is to say if the
asset was not fully depreciated yet.
If the asset was fully depreciated, then no disposal move was generated,
and the asset was set to close without really being disposed of.

Now, a disposal move is generated even for fully depreciated assets.

opw-2391594

closes odoo/enterprise#16422

X-original-commit: d20e5000d6d29e0f9f77aac1ad9ee5bb875adc83
Signed-off-by: William André (wan) <wan@odoo.com>