[FIX] account_reports: EC sales list: make the report 'strict_range'
Before this fix, the selected range of dates wasn't strictly applied; so some lines from previous periods could be taken into account inc ase their account was of a type including initial balance. This is not what we want. v14 did properly, this behavior was introduced by the 14.1 refactoring of this report.

[FIX] account_reports: EC sales list: don't consider UK as a member of EC after the Brexit

[FIX] l10n_be_reports: add missing checks in Belgian's EC sales list implementation

Those were defined in the parent function, but super() is not called in Belgian version of the report.

closes odoo/enterprise#16475

X-original-commit: 50e35531c261eb9b0597a734ef557236c19110d2
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>